add a filter to pkg/zip to extract only necessary binary
