--- pkg/check/binary/testdata/test.txt
@@ -1,1 +1,0 @@
-This is text file for testing
