+++ versionmanager/retriever/tofu/tofuretriever.go
@@ -34,6 +34,8 @@ import (
 	sha256check "github.com/tofuutils/tenv/v2/pkg/check/sha256"
 	"github.com/tofuutils/tenv/v2/pkg/download"
 	"github.com/tofuutils/tenv/v2/pkg/github"
+	"github.com/tofuutils/tenv/v2/pkg/pathfilter"
+	"github.com/tofuutils/tenv/v2/pkg/winbin"
 	"github.com/tofuutils/tenv/v2/pkg/zip"
 	htmlretriever "github.com/tofuutils/tenv/v2/versionmanager/retriever/html"
 )
@@ -114,7 +116,7 @@ func (r TofuRetriever) InstallRelease(versionStr string, targetPath string) erro
 		return err
 	}
 
-	return zip.UnzipToDir(data, targetPath)
+	return zip.UnzipToDir(data, targetPath, pathfilter.NameEqual(winbin.GetBinaryName(config.TofuName)))
 }
 
 func (r TofuRetriever) ListReleases() ([]string, error) {
